Search through all messages when looking for a message
Before this extension would mostly only look if the last sent message
adhered to some criteria, but now it will look through all messages in
the inbox to see if there is any message matching the criteria.

That does mean that searching might be slower, because it will go
through the entire mailhog inbox for each search.

Make sure to purge the mailbox every once in a while to avoid it getting
too slow.